Dummy Data
This is some dummy data to trouble shoot the Trusting Trends code with. I took the actual data (which requires permission from the Christmas Bird Count to access) and wrote over everything, so they are fake Population names, randomly simulated Negative Binomial counts, Years from 1:30 and Hours set to 1. It’s in an RData format, object name “TYB_Dummy”.
 
It's the same length as the actual data used with the code just to try and keep it as consistent as possible for error checking, but obviously you can reduce it to just a few populations so it doesn’t take a long time to run.